+++ samples/openapi3/client/petstore/python/src/petstore_api/paths/fake_test_query_paramters/put/operation.py
@@ -71,7 +71,7 @@ class BaseApi(api_client.Api):
         query_params: RequestQueryParameters.Params = frozendict.frozendict(),
         server_index: typing.Optional[int] = None,
         stream: bool = False,
-        timeout: typing.Optional[typing.Union[int, typing.Tuple]] = None,
+        timeout: typing.Optional[typing.Union[float, typing.Tuple]] = None,
         skip_deserialization: typing_extensions.Literal[False] = ...,
     ) -> response_200.ResponseFor200.response_cls: ...
 
@@ -82,7 +82,7 @@ class BaseApi(api_client.Api):
         query_params: RequestQueryParameters.Params = frozendict.frozendict(),
         server_index: typing.Optional[int] = None,
         stream: bool = False,
-        timeout: typing.Optional[typing.Union[int, typing.Tuple]] = None,
+        timeout: typing.Optional[typing.Union[float, typing.Tuple]] = None,
     ) -> api_response.ApiResponseWithoutDeserialization: ...
 
     @typing.overload
@@ -91,7 +91,7 @@ class BaseApi(api_client.Api):
         query_params: RequestQueryParameters.Params = frozendict.frozendict(),
         server_index: typing.Optional[int] = None,
         stream: bool = False,
-        timeout: typing.Optional[typing.Union[int, typing.Tuple]] = None,
+        timeout: typing.Optional[typing.Union[float, typing.Tuple]] = None,
         skip_deserialization: bool = ...,
     ) -> typing.Union[
         response_200.ResponseFor200.response_cls,
@@ -103,7 +103,7 @@ class BaseApi(api_client.Api):
         query_params: RequestQueryParameters.Params = frozendict.frozendict(),
         server_index: typing.Optional[int] = None,
         stream: bool = False,
-        timeout: typing.Optional[typing.Union[int, typing.Tuple]] = None,
+        timeout: typing.Optional[typing.Union[float, typing.Tuple]] = None,
         skip_deserialization: bool = False,
     ):
         """
