+++ samples/openapi3/client/petstore/python/src/petstore_api/paths/pet_pet_id/delete/operation.py
@@ -91,7 +91,7 @@ class BaseApi(api_client.Api):
         security_index: typing.Optional[int] = None,
         server_index: typing.Optional[int] = None,
         stream: bool = False,
-        timeout: typing.Optional[typing.Union[int, typing.Tuple]] = None,
+        timeout: typing.Optional[typing.Union[float, typing.Tuple]] = None,
         skip_deserialization: typing_extensions.Literal[False] = ...,
     ) -> api_response.ApiResponseWithoutDeserialization: ...
     @typing.overload
@@ -103,7 +103,7 @@ class BaseApi(api_client.Api):
         security_index: typing.Optional[int] = None,
         server_index: typing.Optional[int] = None,
         stream: bool = False,
-        timeout: typing.Optional[typing.Union[int, typing.Tuple]] = None,
+        timeout: typing.Optional[typing.Union[float, typing.Tuple]] = None,
     ) -> api_response.ApiResponseWithoutDeserialization: ...
 
     @typing.overload
@@ -114,7 +114,7 @@ class BaseApi(api_client.Api):
         security_index: typing.Optional[int] = None,
         server_index: typing.Optional[int] = None,
         stream: bool = False,
-        timeout: typing.Optional[typing.Union[int, typing.Tuple]] = None,
+        timeout: typing.Optional[typing.Union[float, typing.Tuple]] = None,
         skip_deserialization: bool = ...,
     ) -> typing.Union[
         api_response.ApiResponseWithoutDeserialization,
@@ -127,7 +127,7 @@ class BaseApi(api_client.Api):
         security_index: typing.Optional[int] = None,
         server_index: typing.Optional[int] = None,
         stream: bool = False,
-        timeout: typing.Optional[typing.Union[int, typing.Tuple]] = None,
+        timeout: typing.Optional[typing.Union[float, typing.Tuple]] = None,
         skip_deserialization: bool = False,
     ):
         """
