+++ samples/openapi3/client/petstore/python/src/petstore_api/paths/pet_pet_id/post/operation.py
@@ -72,7 +72,7 @@ class BaseApi(api_client.Api):
         security_index: typing.Optional[int] = None,
         server_index: typing.Optional[int] = None,
         stream: bool = False,
-        timeout: typing.Optional[typing.Union[int, typing.Tuple]] = None,
+        timeout: typing.Optional[typing.Union[float, typing.Tuple]] = None,
         skip_deserialization: typing_extensions.Literal[False] = ...,
     ) -> api_response.ApiResponseWithoutDeserialization: ...
     @typing.overload
@@ -89,7 +89,7 @@ class BaseApi(api_client.Api):
         security_index: typing.Optional[int] = None,
         server_index: typing.Optional[int] = None,
         stream: bool = False,
-        timeout: typing.Optional[typing.Union[int, typing.Tuple]] = None,
+        timeout: typing.Optional[typing.Union[float, typing.Tuple]] = None,
         skip_deserialization: typing_extensions.Literal[False] = ...,
     ) -> api_response.ApiResponseWithoutDeserialization: ...
 
@@ -108,7 +108,7 @@ class BaseApi(api_client.Api):
         security_index: typing.Optional[int] = None,
         server_index: typing.Optional[int] = None,
         stream: bool = False,
-        timeout: typing.Optional[typing.Union[int, typing.Tuple]] = None,
+        timeout: typing.Optional[typing.Union[float, typing.Tuple]] = None,
     ) -> api_response.ApiResponseWithoutDeserialization: ...
 
     @typing.overload
@@ -125,7 +125,7 @@ class BaseApi(api_client.Api):
         security_index: typing.Optional[int] = None,
         server_index: typing.Optional[int] = None,
         stream: bool = False,
-        timeout: typing.Optional[typing.Union[int, typing.Tuple]] = None,
+        timeout: typing.Optional[typing.Union[float, typing.Tuple]] = None,
         skip_deserialization: bool = ...,
     ) -> typing.Union[
         api_response.ApiResponseWithoutDeserialization,
@@ -144,7 +144,7 @@ class BaseApi(api_client.Api):
         security_index: typing.Optional[int] = None,
         server_index: typing.Optional[int] = None,
         stream: bool = False,
-        timeout: typing.Optional[typing.Union[int, typing.Tuple]] = None,
+        timeout: typing.Optional[typing.Union[float, typing.Tuple]] = None,
         skip_deserialization: bool = False,
     ):
         """
